minisite deletion works
